Korjaa validointi EU-rahoitukselle, suunnittelusopimukselle, kielitiedoille ja vähäiselle menettelylle, ja tee puuttuvat testin BE:iin.
