+++ backend/src/projekti/validator/projektiValidator.ts
@@ -1,7 +1,6 @@
 import { DBProjekti, UudelleenKuulutus } from "../../database/model";
 import {
   KayttajaTyyppi,
-  KuulutusJulkaisuTila,
   Projekti,
   ProjektiTyyppi,
   Status,
@@ -93,9 +92,6 @@ function validateKielivalinta(dbProjekti: DBProjekti, input: TallennaProjektiInp
   }
 }
 
-/**
- * Validoi, että suunnittelusopimusta ei poisteta tai lisätä sen jälkeen kun aloituskuulutusjulkaisu on hyväksynnässä tai hyväksytty
- */
 function validateSuunnitteluSopimus(dbProjekti: DBProjekti, input: TallennaProjektiInput) {
   const suunnitteluSopimusAfterSaving: boolean =
     !!input.suunnitteluSopimus || !!(input.suunnitteluSopimus === undefined && dbProjekti.suunnitteluSopimus);
@@ -116,19 +112,25 @@ function validateSuunnitteluSopimus(dbProjekti: DBProjekti, input: TallennaProje
     (input.suunnitteluSopimus === null && !!dbProjekti.suunnitteluSopimus) ||
     (!!input.suunnitteluSopimus && !dbProjekti.suunnitteluSopimus);
 
-  const aloituskuulutusjulkaisuja = dbProjekti?.aloitusKuulutusJulkaisut?.length;
-  if (!aloituskuulutusjulkaisuja || aloituskuulutusjulkaisuja < 1) {
-    return;
-  } // Lista voi olla myos olemassa, mutta tyhja, jos kuulutus on esim palautettu muokattavaksi
+  const allowedToChangeSuunnittelusopimus = isAllowedToChangeSuunnittelusopimus(dbProjekti);
 
-  const latestAloituskuulutusJulkaisuTila = dbProjekti?.aloitusKuulutusJulkaisut?.[dbProjekti.aloitusKuulutusJulkaisut.length - 1].tila;
-  const isLatestJulkaisuPendingApprovalOrApproved =
-    !!latestAloituskuulutusJulkaisuTila &&
-    [KuulutusJulkaisuTila.HYVAKSYTTY, KuulutusJulkaisuTila.ODOTTAA_HYVAKSYNTAA].includes(latestAloituskuulutusJulkaisuTila);
-
-  if (isSuunnitteluSopimusAddedOrDeleted && isLatestJulkaisuPendingApprovalOrApproved) {
+  if (isSuunnitteluSopimusAddedOrDeleted && !allowedToChangeSuunnittelusopimus) {
     throw new IllegalArgumentError(
-      "Suunnittelusopimuksen olemassaoloa ei voi muuttaa, jos aloituskuulutus on jo julkaistu tai se odottaa hyväksyntää!"
+      "EU-rahoituksen olemassaoloa ei voi muuttaa, jos ensimmäinen HASSUssa tehty vaihe ei ole muokkaustilassa, tai jos ensimmäinen vaihe on suunnittelu, yhtäkään kutsua ei ole tehty."
+    );
+  }
+}
+
+function validateEuRahoitus(dbProjekti: DBProjekti, input: TallennaProjektiInput) {
+  const isEuSopimusAddedOrDeleted =
+    ((input.euRahoitus === null || input.euRahoitus === false) && !!dbProjekti.euRahoitus) ||
+    (!!input.euRahoitus && !dbProjekti.euRahoitus);
+
+  const allowedToChangeEuSopimus = isAllowedToChangeEuRahoitus(dbProjekti);
+
+  if (isEuSopimusAddedOrDeleted && !allowedToChangeEuSopimus) {
+    throw new IllegalArgumentError(
+      "EU-rahoituksen olemassaoloa ei voi muuttaa, jos ensimmäinen HASSUssa tehty vaihe ei ole muokkaustilassa, tai jos ensimmäinen vaihe on suunnittelu, yhtäkään kutsua ei ole tehty."
     );
   }
 }
@@ -137,6 +139,7 @@ export async function validateTallennaProjekti(projekti: DBProjekti, input: Tall
   requirePermissionMuokkaa(projekti);
   const apiProjekti = await projektiAdapter.adaptProjekti(projekti);
   validateKielivalinta(projekti, input);
+  validateEuRahoitus(projekti, input);
   validateKasittelynTila(projekti, apiProjekti, input);
   validateVarahenkiloModifyPermissions(projekti, input);
   validateSuunnitteluSopimus(projekti, input);
@@ -197,9 +200,6 @@ export function validatePaivitaPerustiedot(projekti: DBProjekti, input: Vuorovai
   }
 }
 
-/**
- * Validoi, että vahainenMenettely-tietoa ei muokata sen jälkeen kun aloituskuulutusjulkaisu on hyväksynnässä tai hyväksytty
- */
 function validateVahainenMenettely(dbProjekti: DBProjekti, input: TallennaProjektiInput) {
   const vahainenMenettelyAfterSaving: boolean =
     input.vahainenMenettely === true || (input.vahainenMenettely === undefined && dbProjekti.vahainenMenettely === true);
@@ -210,23 +210,18 @@ function validateVahainenMenettely(dbProjekti: DBProjekti, input: TallennaProjek
     throw new IllegalArgumentError("Projekteilla, joihin sovelletaan vähäistä menettelyä, ei voi olla suunnittelusopimusta.");
   }
 
-  const aloituskuulutusjulkaisuja = dbProjekti?.aloitusKuulutusJulkaisut?.length;
-  if (!aloituskuulutusjulkaisuja || aloituskuulutusjulkaisuja < 1) {
-    return;
-  } // Lista voi olla myos olemassa, mutta tyhja, jos kuulutus on esim palautettu muokattavaksi
-
   const isVahainenMenettelyValueChanged =
     typeof input.vahainenMenettely === "boolean" && !!input.vahainenMenettely !== !!dbProjekti.vahainenMenettely;
 
   const allowedToChangeVahainenMenettely = isAllowedToChangeVahainenMenettely(dbProjekti);
   if (isVahainenMenettelyValueChanged && !allowedToChangeVahainenMenettely) {
     throw new IllegalArgumentError(
-      "Vähäinen menettely -tietoa ei voi muuttaa, jos aloituskuulutus on jo julkaistu tai se odottaa hyväksyntää!"
+      "EU-rahoituksen olemassaoloa ei voi muuttaa, jos ensimmäinen HASSUssa tehty vaihe ei ole muokkaustilassa, tai jos ensimmäinen vaihe on suunnittelu, yhtäkään kutsua ei ole tehty."
     );
   }
 }
 
-function isAllowedToChangeVahainenMenettely(dbProjekti: DBProjekti) {
+function noJulkaisuIsInReadStateDBProjekti(dbProjekti: DBProjekti) {
   return noJulkaisuOrKutsuIsInReadState({
     aloitusKuulutusMuokkausTila: dbProjekti.aloitusKuulutus
       ? adaptMuokkausTila(dbProjekti.aloitusKuulutus, dbProjekti.aloitusKuulutusJulkaisut)
@@ -245,6 +240,27 @@ function isAllowedToChangeVahainenMenettely(dbProjekti: DBProjekti) {
       ? adaptMuokkausTila(dbProjekti.jatkoPaatos2Vaihe, dbProjekti.jatkoPaatos2VaiheJulkaisut)
       : undefined,
   });
+}
+
+function thereAreNoVuorovaikutusKierrosJulkaisutThatAreNotMigroitu(dbProjekti: DBProjekti) {
+  return (
+    !dbProjekti.vuorovaikutusKierrosJulkaisut ||
+    !dbProjekti.vuorovaikutusKierrosJulkaisut.length ||
+    (dbProjekti.vuorovaikutusKierrosJulkaisut.length === 1 &&
+      dbProjekti.vuorovaikutusKierrosJulkaisut[0].tila === VuorovaikutusKierrosTila.MIGROITU)
+  );
+}
+
+function isAllowedToChangeVahainenMenettely(dbProjekti: DBProjekti) {
+  return thereAreNoVuorovaikutusKierrosJulkaisutThatAreNotMigroitu(dbProjekti) && noJulkaisuIsInReadStateDBProjekti(dbProjekti);
+}
+
+function isAllowedToChangeSuunnittelusopimus(dbProjekti: DBProjekti) {
+  return thereAreNoVuorovaikutusKierrosJulkaisutThatAreNotMigroitu(dbProjekti) && noJulkaisuIsInReadStateDBProjekti(dbProjekti);
+}
+
+function isAllowedToChangeEuRahoitus(dbProjekti: DBProjekti) {
+  return thereAreNoVuorovaikutusKierrosJulkaisutThatAreNotMigroitu(dbProjekti) && noJulkaisuIsInReadStateDBProjekti(dbProjekti);
 }
 
 function validateVuorovaikutuskierrokset(projekti: DBProjekti, input: TallennaProjektiInput) {
