+++ backend/test/projekti/validator/projektiValidator/vahainenMenettely.test.ts
@@ -1,33 +1,21 @@
 import { describe, it } from "mocha";
-import { ProjektiFixture } from "../../fixture/projektiFixture";
+import { ProjektiFixture } from "../../../fixture/projektiFixture";
 import * as sinon from "sinon";
-import { UserFixture } from "../../fixture/userFixture";
-import { userService } from "../../../src/user";
-import { personSearch } from "../../../src/personSearch/personSearchClient";
-import { PersonSearchFixture } from "../../personSearch/lambda/personSearchFixture";
-import { Kayttajas } from "../../../src/personSearch/kayttajas";
-import { validateTallennaProjekti } from "../../../src/projekti/validator/projektiValidator";
-import {
-  AineistoTila,
-  ELY,
-  KayttajaTyyppi,
-  Kieli,
-  KuulutusJulkaisuTila,
-  ProjektiKayttajaInput,
-  TallennaProjektiInput,
-  VuorovaikutusKierrosTila,
-} from "../../../../common/graphql/apiModel";
-import assert from "assert";
-import { kategorisoimattomatId } from "../../../../common/aineistoKategoriat";
-import { Aineisto, DBProjekti, UudelleenkuulutusTila, Velho } from "../../../src/database/model";
-import { IllegalArgumentError } from "../../../src/error/IllegalArgumentError";
-import { assertIsDefined } from "../../../src/util/assertions";
+import { UserFixture } from "../../../fixture/userFixture";
+import { userService } from "../../../../src/user";
+import { personSearch } from "../../../../src/personSearch/personSearchClient";
+import { PersonSearchFixture } from "../../../personSearch/lambda/personSearchFixture";
+import { Kayttajas } from "../../../../src/personSearch/kayttajas";
+import { validateTallennaProjekti } from "../../../../src/projekti/validator/projektiValidator";
+import { Kieli, KuulutusJulkaisuTila, TallennaProjektiInput, VuorovaikutusKierrosTila } from "../../../../../common/graphql/apiModel";
+import { DBProjekti, UudelleenkuulutusTila } from "../../../../src/database/model";
+import { IllegalArgumentError } from "../../../../src/error/IllegalArgumentError";
 import { expect } from "chai";
 
 const ELY_UID = "A1";
 const VAYLA_UID = "A2";
 
-describe("projektiValidator", () => {
+describe("projektiValidator (vahainenMenettelyValidator)", () => {
   let fixture: ProjektiFixture;
   const userFixture = new UserFixture(userService);
 
@@ -46,267 +34,6 @@ describe("projektiValidator", () => {
   afterEach(() => {
     userFixture.logout();
     sinon.restore();
-  });
-
-  it("Muokkaaminen vaatii oikeudet", async () => {
-    userFixture.logout();
-    const projekti = fixture.dbProjekti2();
-    await expect(validateTallennaProjekti(projekti, { oid: projekti.oid, versio: projekti.versio })).to.eventually.rejectedWith(
-      "Väylä-kirjautuminen puuttuu"
-    );
-
-    userFixture.loginAs(UserFixture.pekkaProjari);
-    // Should validate just fine
-    await validateTallennaProjekti(projekti, { oid: projekti.oid, versio: projekti.versio });
-  });
-
-  it("Vain omistaja voi lisätä varahenkilöitä", async () => {
-    await doTestModifyVarahenkiloAsNonOwner(undefined, KayttajaTyyppi.VARAHENKILO);
-  });
-
-  it("Vain omistaja voi poistaa varahenkilöitä", async () => {
-    await doTestModifyVarahenkiloAsNonOwner(KayttajaTyyppi.VARAHENKILO, undefined);
-  });
-
-  async function doTestModifyVarahenkiloAsNonOwner(initialType: KayttajaTyyppi | undefined, targetType: KayttajaTyyppi | undefined) {
-    userFixture.logout();
-    const projekti = fixture.dbProjekti2();
-    projekti.kayttoOikeudet.push(fixture.kunnanYhteysHenkiloDBVaylaUser());
-
-    const varahenkiloKayttajaTunnus = fixture.mattiMeikalainenDBVaylaUser().kayttajatunnus;
-
-    // Alkutilanne: mattiMeikalainen ei ole varahenkilo
-    const kayttaja = projekti.kayttoOikeudet.filter((user) => user.kayttajatunnus == varahenkiloKayttajaTunnus).pop();
-    assert(kayttaja);
-    kayttaja.tyyppi = initialType;
-    kayttaja.muokattavissa = true;
-
-    // Yritä muuttaa mattiMeikalainen varahenkiloksi
-    const kayttoOikeudetInput = projekti.kayttoOikeudet.map((user) => {
-      assertIsDefined(user.puhelinnumero);
-      const kayttoOikeus: ProjektiKayttajaInput = {
-        kayttajatunnus: user.kayttajatunnus,
-        yleinenYhteystieto: user.yleinenYhteystieto,
-        puhelinnumero: user.puhelinnumero,
-        tyyppi: user.kayttajatunnus == varahenkiloKayttajaTunnus ? targetType : user.tyyppi,
-      };
-      return kayttoOikeus;
-    });
-
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      kayttoOikeudet: kayttoOikeudetInput,
-    };
-
-    // Ei sallittua muuttaa varahenkilöitä jos ei ole omistaja
-    userFixture.loginAs(UserFixture.kunnanYhteysHenkiloProjektiKayttaja);
-    await expect(validateTallennaProjekti(projekti, input)).to.eventually.rejectedWith("Et ole projektin omistaja");
-
-    // Sallittua projektin projektipäällikölle
-    userFixture.loginAs(UserFixture.pekkaProjari);
-    await validateTallennaProjekti(projekti, input);
-  }
-
-  it("Nähtävilläolovaiheen kuulutustietoja ei voi tallentaa ennen aineistojen tallentamista", async () => {
-    userFixture.loginAs(UserFixture.mattiMeikalainen);
-    const projekti = fixture.dbProjektiLackingNahtavillaoloVaihe();
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      nahtavillaoloVaihe: { hankkeenKuvaus: { SUOMI: "Kuvaus" } },
-    };
-    expect(validateTallennaProjekti(projekti, input)).to.eventually.rejectedWith(
-      "Nähtävilläolovaiheen aineistoja ei ole vielä tallennettu tai niiden joukossa on kategorisoimattomia."
-    );
-  });
-
-  it("Nähtävilläolovaiheen kuulutustietoja ei voi tallentaa, jos aineistoja kategorisoimatta", async () => {
-    userFixture.loginAs(UserFixture.mattiMeikalainen);
-    const projekti = fixture.dbProjektiLackingNahtavillaoloVaihe();
-    projekti.nahtavillaoloVaihe = {
-      id: 1,
-      aineistoNahtavilla: [
-        {
-          dokumenttiOid: "11",
-          jarjestys: 1,
-          kategoriaId: kategorisoimattomatId,
-          nimi: "T113 TS Esite.txt",
-          tiedosto: "/nahtavillaolo/1/T113 TS Esite.txt",
-          tila: AineistoTila.VALMIS,
-          tuotu: "***unittest***",
-        },
-      ],
-    };
-
-    const inputWithKuulutusData: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      nahtavillaoloVaihe: { hankkeenKuvaus: { SUOMI: "Kuvaus" } },
-    };
-    expect(validateTallennaProjekti(projekti, inputWithKuulutusData)).to.eventually.rejectedWith(
-      "Nähtävilläolovaiheen aineistoja ei ole vielä tallennettu tai niiden joukossa on kategorisoimattomia."
-    );
-    projekti.nahtavillaoloVaihe.aineistoNahtavilla?.forEach((aineisto) => {
-      aineisto.kategoriaId = undefined;
-    });
-    expect(validateTallennaProjekti(projekti, inputWithKuulutusData)).to.eventually.rejectedWith(
-      "Nähtävilläolovaiheen aineistoja ei ole vielä tallennettu tai niiden joukossa on kategorisoimattomia."
-    );
-    // Tallennus on OK kun kaikilla aineistoilla on kategoriat
-    projekti.nahtavillaoloVaihe.aineistoNahtavilla?.forEach((aineisto) => {
-      aineisto.kategoriaId = "osa_a";
-    });
-    await validateTallennaProjekti(projekti, inputWithKuulutusData);
-  });
-
-  it("Hyväksymispäätösvaiheen kuulutustietoja ei voi tallentaa ennen aineistojen tallentamista", async () => {
-    userFixture.loginAs(UserFixture.mattiMeikalainen);
-    const projekti = fixture.dbProjektiHyvaksymisMenettelyssa();
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      hyvaksymisPaatosVaihe: { kuulutusPaiva: "2023-01-01" },
-    };
-    expect(validateTallennaProjekti(projekti, input)).to.eventually.be.rejectedWith(
-      "hyvaksymisPaatosVaihe aineistoja ei ole vielä tallennettu tai niiden joukossa on kategorisoimattomia."
-    );
-  });
-
-  it("Hyväksymispäätösvaiheen kuulutustietoja ei voi tallentaa, jos aineistoja kategorisoimatta", async () => {
-    userFixture.loginAs(UserFixture.mattiMeikalainen);
-    const aineisto: Aineisto = {
-      dokumenttiOid: "11",
-      jarjestys: 1,
-      kategoriaId: kategorisoimattomatId,
-      nimi: "T113 TS Esite.txt",
-      tiedosto: "/hyvaksymispaatos/1/T113 TS Esite.txt",
-      tila: AineistoTila.VALMIS,
-      tuotu: "***unittest***",
-    };
-    const projekti = fixture.dbProjektiHyvaksymisMenettelyssa();
-    projekti.hyvaksymisPaatosVaihe = {
-      id: 1,
-      aineistoNahtavilla: [aineisto],
-      hyvaksymisPaatos: [aineisto],
-    };
-
-    const inputWithKuulutusData: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      hyvaksymisPaatosVaihe: { kuulutusPaiva: "2022-01-01" },
-    };
-    expect(validateTallennaProjekti(projekti, inputWithKuulutusData)).to.eventually.rejectedWith(
-      "hyvaksymisPaatosVaihe aineistoja ei ole vielä tallennettu tai niiden joukossa on kategorisoimattomia."
-    );
-    projekti.hyvaksymisPaatosVaihe.aineistoNahtavilla?.forEach((aineisto) => {
-      aineisto.kategoriaId = undefined;
-    });
-    expect(validateTallennaProjekti(projekti, inputWithKuulutusData)).to.eventually.rejectedWith(
-      "hyvaksymisPaatosVaihe aineistoja ei ole vielä tallennettu tai niiden joukossa on kategorisoimattomia."
-    );
-    // Tallennus on OK kun kaikilla aineistoilla on kategoriat
-    projekti.hyvaksymisPaatosVaihe.aineistoNahtavilla?.forEach((aineisto) => {
-      aineisto.kategoriaId = "osa_a";
-    });
-    await validateTallennaProjekti(projekti, inputWithKuulutusData);
-  });
-
-  it("Hyväksymispäätösvaiheen kuulutustietoja ei voi tallentaa, jos päätös tallentamatta", async () => {
-    userFixture.loginAs(UserFixture.mattiMeikalainen);
-    const aineisto: Aineisto = {
-      dokumenttiOid: "11",
-      jarjestys: 1,
-      kategoriaId: "osa_a",
-      nimi: "T113 TS Esite.txt",
-      tiedosto: "/hyvaksymispaatos/1/T113 TS Esite.txt",
-      tila: AineistoTila.VALMIS,
-      tuotu: "***unittest***",
-    };
-    const projekti = fixture.dbProjektiHyvaksymisMenettelyssa();
-    projekti.hyvaksymisPaatosVaihe = {
-      id: 1,
-      aineistoNahtavilla: [aineisto],
-    };
-
-    const inputWithKuulutusData: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      hyvaksymisPaatosVaihe: { kuulutusPaiva: "2022-01-01" },
-    };
-    expect(validateTallennaProjekti(projekti, inputWithKuulutusData)).to.eventually.rejectedWith(
-      "hyvaksymisPaatosVaihe aineistoja ei ole vielä tallennettu tai niiden joukossa on kategorisoimattomia."
-    );
-    // Tallennus on OK kun hyväksymispäätös on asetettu
-    projekti.hyvaksymisPaatosVaihe.hyvaksymisPaatos = [aineisto];
-    await validateTallennaProjekti(projekti, inputWithKuulutusData);
-  });
-
-  it("elyOrganisaatio tiedon voi tallettaa kayttajalle, jolla organisaatio on asetettu 'ELY':ksi", async () => {
-    userFixture.loginAs(UserFixture.pekkaProjari);
-    const projekti = fixture.dbProjektiHyvaksymisMenettelyssa();
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      kayttoOikeudet: [{ kayttajatunnus: ELY_UID, puhelinnumero: "0441231234", elyOrganisaatio: ELY.HAME_ELY }],
-    };
-    // Tallennus on OK kun tallennettavan käyttäjän organisaatio on 'ELY'
-    await expect(validateTallennaProjekti(projekti, input)).to.eventually.be.fulfilled;
-  });
-
-  it("elyOrganisaatio tietoa ei voi tallettaa kayttajalle, jolla organisaatio ei ole 'ELY'", async () => {
-    userFixture.loginAs(UserFixture.pekkaProjari);
-    const projekti = fixture.dbProjektiHyvaksymisMenettelyssa();
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      kayttoOikeudet: [{ kayttajatunnus: VAYLA_UID, puhelinnumero: "0441231234", elyOrganisaatio: ELY.HAME_ELY }],
-    };
-    // Tallennus epäonnistuu kun tallennettavan käyttäjän organisaatio ei ole 'ELY'
-    await expect(validateTallennaProjekti(projekti, input)).to.eventually.be.rejectedWith(IllegalArgumentError);
-  });
-
-  it("ei anna asettaa vähäiseen menettelyyn suunnittelusopimuksellista projektia", async () => {
-    userFixture.loginAs(UserFixture.pekkaProjari);
-    const { aloitusKuulutus, ...projekti } = fixture.dbProjekti1(); // Tällä on suunnittelusopimus
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      vahainenMenettely: true,
-    };
-    await expect(validateTallennaProjekti(projekti, input)).to.eventually.be.rejectedWith(IllegalArgumentError);
-  });
-
-  it("ei anna asettaa vähäiseen menettelyyn projektia jonka asettaa myös suunnittelusopimukselliseksi", async () => {
-    userFixture.loginAs(UserFixture.pekkaProjari);
-    const { aloitusKuulutus, suunnitteluSopimus, ...projekti } = fixture.dbProjekti1();
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      vahainenMenettely: true,
-      suunnitteluSopimus: {
-        yhteysHenkilo: "plop",
-        kunta: 1,
-        logo: "jotain.png",
-      },
-    };
-    await expect(validateTallennaProjekti(projekti, input)).to.eventually.be.rejectedWith(IllegalArgumentError);
-  });
-
-  it("ei anna asettaa suunnittelusopimusta projektille, jossa sovelletaan vähäistä menettelyä", async () => {
-    userFixture.loginAs(UserFixture.pekkaProjari);
-    const { aloitusKuulutus, suunnitteluSopimus, ...projekti } = fixture.dbProjekti1();
-    projekti.vahainenMenettely = true;
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      suunnitteluSopimus: {
-        yhteysHenkilo: "plop",
-        kunta: 1,
-        logo: "jotain.png",
-      },
-    };
-    await expect(validateTallennaProjekti(projekti, input)).to.eventually.be.rejectedWith(IllegalArgumentError);
   });
 
   it("ei anna asettaa vähäistä menettelyä projektille, jolla on julkaistu aloituskuulutus eikä muokkaustilaista uudelleenkuulutusta", async () => {
@@ -816,82 +543,5 @@ describe("projektiValidator", () => {
       vahainenMenettely: true,
     };
     await expect(validateTallennaProjekti(migroituProjekti, input)).to.eventually.be.rejectedWith(IllegalArgumentError);
-  });
-
-  it("antaa muuttaa kielitietoja projektille, jolla ei ole aloituskuulutusjulkaisua", async () => {
-    userFixture.loginAs(UserFixture.pekkaProjari);
-    const { suunnitteluSopimus, nahtavillaoloVaihe, hyvaksymisPaatosVaihe, aloitusKuulutusJulkaisut, ...projekti } = fixture.dbProjekti2();
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      kielitiedot: {
-        ensisijainenKieli: Kieli.SUOMI,
-        toissijainenKieli: null,
-      },
-    };
-    await expect(await validateTallennaProjekti(projekti, input)).to.eql(undefined);
-  });
-
-  it("antaa muuttaa kielitietoja projektille, jolla on migroitu aloituskuulutusjulkaisu", async () => {
-    userFixture.loginAs(UserFixture.pekkaProjari);
-    const { suunnitteluSopimus, nahtavillaoloVaihe, hyvaksymisPaatosVaihe, aloitusKuulutusJulkaisut, aloitusKuulutus, ...projekti } =
-      fixture.dbProjekti2();
-    const testiProjekti: DBProjekti = {
-      ...projekti,
-      aloitusKuulutusJulkaisut: [
-        { id: 1, tila: KuulutusJulkaisuTila.MIGROITU, yhteystiedot: [], kielitiedot: projekti.kielitiedot, velho: projekti.velho as Velho },
-      ],
-    };
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      kielitiedot: {
-        ensisijainenKieli: Kieli.SUOMI,
-        toissijainenKieli: null,
-      },
-    };
-    await expect(await validateTallennaProjekti(testiProjekti, input)).to.eql(undefined);
-  });
-
-  it("ei anna muuttaa kielitietoja projektille, jolla on julkaistu aloituskuulutus eikä muokkaustilaista uudelleenkuulutusta", async () => {
-    userFixture.loginAs(UserFixture.pekkaProjari);
-    const { suunnitteluSopimus, nahtavillaoloVaihe, hyvaksymisPaatosVaihe, ...projekti } = fixture.dbProjekti2();
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      kielitiedot: {
-        ensisijainenKieli: Kieli.SUOMI,
-        toissijainenKieli: null,
-      },
-    };
-    await expect(validateTallennaProjekti(projekti, input)).to.eventually.be.rejectedWith(IllegalArgumentError);
-  });
-
-  it("ei anna muuttaa kielitietoja projektille, vaikka sillä olisi on muokattavaksi palautettu aloituskuulutuksen uudelleenkuulutus", async () => {
-    userFixture.loginAs(UserFixture.pekkaProjari);
-    const { suunnitteluSopimus, nahtavillaoloVaihe, hyvaksymisPaatosVaihe, ...projekti } = fixture.dbProjekti2();
-    const input: TallennaProjektiInput = {
-      oid: projekti.oid,
-      versio: projekti.versio,
-      kielitiedot: {
-        ensisijainenKieli: Kieli.SUOMI,
-        toissijainenKieli: null,
-      },
-    };
-    projekti.aloitusKuulutus = {
-      ...projekti.aloitusKuulutus,
-      id: projekti.aloitusKuulutus?.id || 1,
-      uudelleenKuulutus: {
-        alkuperainenHyvaksymisPaiva: "2023-06-15",
-        selosteKuulutukselle: {
-          SUOMI: "sdgsdfg",
-        },
-        selosteLahetekirjeeseen: {
-          SUOMI: "afgafhg",
-        },
-        tila: UudelleenkuulutusTila.JULKAISTU_PERUUTETTU,
-      },
-    };
-    await expect(validateTallennaProjekti(projekti, input)).to.eventually.be.rejectedWith(IllegalArgumentError);
   });
 });
