+++ src/components/projekti/ProjektiSunnittelusopimusTiedot.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement, useEffect, useState } from "react";
-import { KuulutusJulkaisuTila, Projekti, ProjektiTyyppi, SuunnitteluSopimusInput } from "@services/api";
+import { Projekti, ProjektiTyyppi, SuunnitteluSopimusInput } from "@services/api";
 import RadioButton from "@components/form/RadioButton";
 import Select, { SelectOption } from "@components/form/Select";
 import { Controller, useFormContext } from "react-hook-form";
@@ -15,13 +15,14 @@ import useTranslation from "next-translate/useTranslation";
 import { kuntametadata } from "../../../common/kuntametadata";
 import { formatNimi } from "../../util/userUtil";
 import Notification, { NotificationType } from "@components/notification/Notification";
+import { isAllowedToChangeSuunnittelusopimus } from "common/util/operationValidators";
 
 interface Props {
   projekti?: Projekti | null;
-  kuntalista?: string[];
+  formDisabled?: boolean;
 }
 
-export default function ProjektiPerustiedot({ projekti }: Props): ReactElement {
+export default function ProjektiPerustiedot({ formDisabled, projekti }: Props): ReactElement {
   const {
     register,
     formState: { errors },
@@ -34,12 +35,6 @@ export default function ProjektiPerustiedot({ projekti }: Props): ReactElement {
   const [logoUrl, setLogoUrl] = useState<string | undefined>(undefined);
   const [kuntaOptions, setKuntaOptions] = useState<SelectOption[]>([]);
   const { lang } = useTranslation();
-
-  const disabled = !!(
-    projekti?.aloitusKuulutusJulkaisu &&
-    projekti?.aloitusKuulutusJulkaisu.tila &&
-    [KuulutusJulkaisuTila.HYVAKSYTTY, KuulutusJulkaisuTila.ODOTTAA_HYVAKSYNTAA].includes(projekti.aloitusKuulutusJulkaisu.tila)
-  );
 
   const hide = projekti?.velho.tyyppi === ProjektiTyyppi.RATA || projekti?.velho.tyyppi === ProjektiTyyppi.YLEINEN;
 
@@ -58,9 +53,13 @@ export default function ProjektiPerustiedot({ projekti }: Props): ReactElement {
     setLogoUrl(projekti?.suunnitteluSopimus?.logo || undefined);
   }, [projekti, setHasSuunnitteluSopimus, setLogoUrl]);
 
-  if (!kuntaOptions || kuntaOptions.length == 0 || hide) {
+  if (!kuntaOptions || kuntaOptions.length == 0 || hide || !projekti) {
     return <></>;
   }
+
+  const vahainenMenettelyCanBeChanged = isAllowedToChangeSuunnittelusopimus(projekti);
+
+  const disabled = formDisabled || !vahainenMenettelyCanBeChanged;
 
   return (
     <Section smallGaps>
@@ -116,6 +115,7 @@ export default function ProjektiPerustiedot({ projekti }: Props): ReactElement {
                 }
                 addEmptyOption
                 error={(errors as any).suunnitteluSopimus?.yhteysHenkilo}
+                disabled={formDisabled}
                 {...register("suunnitteluSopimus.yhteysHenkilo", { shouldUnregister: true })}
               />
               <Select
@@ -123,6 +123,7 @@ export default function ProjektiPerustiedot({ projekti }: Props): ReactElement {
                 label="Kunta *"
                 options={kuntaOptions ? kuntaOptions : [{ label: "", value: "" }]}
                 error={(errors as any).suunnitteluSopimus?.kunta}
+                disabled={formDisabled}
                 {...register("suunnitteluSopimus.kunta", { shouldUnregister: true })}
               />
             </HassuGrid>
@@ -138,6 +139,7 @@ export default function ProjektiPerustiedot({ projekti }: Props): ReactElement {
                       <IconButton
                         name="suunnittelusopimus_logo_trash_button"
                         icon="trash"
+                        disabled={formDisabled}
                         onClick={() => {
                           setLogoUrl(undefined);
                           // @ts-ignore
@@ -166,6 +168,7 @@ export default function ProjektiPerustiedot({ projekti }: Props): ReactElement {
                         field.onChange(logoTiedosto);
                       }
                     }}
+                    disabled={formDisabled}
                   />
                 )
               }
@@ -178,8 +181,8 @@ export default function ProjektiPerustiedot({ projekti }: Props): ReactElement {
         </SectionContent>
       )}
       <p>
-        Valintaan voi vaikuttaa aloituskuulutuksen hyväksymiseen saakka, jonka jälkeen valinta lukittuu. Kunnan edustaja on mahdollista
-        vaihtaa myös prosessin aikana.
+        Valintaan voi vaikuttaa aloituskuulutuksen tekemisen saakka, jonka jälkeen valinta lukittuu. Kunnan edustaja on mahdollista vaihtaa
+        myös prosessin aikana.
       </p>
     </Section>
   );
